split onnx code to new file
