+++ deploy/paddle2onnx/readme.md
@@ -55,12 +55,10 @@ python3.7 -m pip install onnxruntime==1.9.0
 接下来就可以用 ONNX 格式模型进行预测，其用法与paddle 预测模型类似
 执行如下命令：
 ```bash
-python3.7 tools/predict.py \
+python3.7 tools/predict_onnx.py \
 --input_file data/example.avi \
---config configs/recognition/pptsn/pptsn_k400_frames.yaml \
---onnx_model_file=./inference/ppTSN/ppTSN.onnx \
---use_gpu=Flase \
---use_onnx=True
+--config configs/recognition/pptsn/pptsn_k400_videos.yaml \
+--model_file=./inference/ppTSN/ppTSN.onnx
 ```
 
 结果如下：
